+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-01-14
+### Changed
+- Split code onto files.
+
+### Added
+- Records validation.
+- Records number.
+
 ## [1.0.2] - 2023-08-03
 ### Fixed
 - Skip unfinished lines.